move DummyExtractor into testing utils
